CU-24wgdhq countryの部分をcityに変更した
