+++ backend/app/Http/Requests/Cities/UpdateCountryRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Countries;
+namespace App\Http\Requests\Cities;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateFavoriteCountryRequest extends FormRequest
+class UpdateCityRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
