+++ backend/app/Models/FavoriteCity.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class FavoriteCountry extends Model
+class FavoriteCity extends Model
 {
     use HasFactory;
 }
