+++ backend/database/factories/Cities/CityFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories\Countries;
+namespace Database\Factories\Cities;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class CountryFactory extends Factory
+class CityFactory extends Factory
 {
     /**
      * Define the model's default state.
