+++ backend/database/factories/Cities/FavoriteCityFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories\Countries;
+namespace Database\Factories\Cities;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class FavoriteCountryFactory extends Factory
+class FavoriteCityFactory extends Factory
 {
     /**
      * Define the model's default state.
